feat: polish auth endpoints
