+++ .env.example
@@ -1,14 +1,16 @@
 # Mongo DB
-DB_CONNECTION=connection_string
-PORT = 3000
+DB_CONNECTION=
+PORT=3000
 
 # Google backend Token
 SERVICE_ACCOUNT_KEY='{
   SERVICE_ACCOUNT_KEY
 }'
-TOKEN_URI=TOKEN_URI
 
-# Google Id Token
-GOOGLE_CLIENT_ID= GOOGLE_CLIENT_ID
-GOOGLE_CLIENT_SECRET= GOOGLE_CLIENT_SECRET
-GOOGLE_REDIRECT_URI= GOOGLE_REDIRECT_URI
+# Google Auth
+GOOGLE_CLIENT_ID=
+GOOGLE_CLIENT_SECRET=
+GOOGLE_REDIRECT_URI=
+TOKEN_URI=
+
+API_URL=
