+++ package.json
@@ -23,7 +23,8 @@
     "google-auth-library": "^9.11.0",
     "jwks-rsa": "^3.1.0",
     "mongoose": "^8.4.1",
-    "nodemon": "^3.1.3"
+    "nodemon": "^3.1.3",
+    "query-string": "^9.0.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.4.0",
