+++ src/middleware/auth.middleware.js
@@ -1,10 +1,14 @@
 import axios from "axios";
 import dotenv from "dotenv";
-import serviceAccount from "../constants/serviceAccount.js";
 
 dotenv.config();
 
 const checkGoogleAccessToken = async (req, res, next) => {
+  if (process.env.DEV_MODE === "true") {
+    next();
+    return;
+  }
+
   const authorizationHeader = req.headers["authorization"];
 
   if (!authorizationHeader) {
@@ -14,14 +18,22 @@ const checkGoogleAccessToken = async (req, res, next) => {
   const token = authorizationHeader.split(" ")[1];
 
   try {
-    const response = await axios.get(`${process.env.TOKEN_URI}${token}`);
+    const uri = process.env.TOKEN_URI;
+
+    const response = await axios.get(uri, {
+      params: {
+        access_token: token,
+      },
+    });
+
     const payload = response.data;
 
-    if (payload.aud !== serviceAccount.client_id) {
+    if (payload.aud !== process.env.GOOGLE_CLIENT_ID) {
       throw new Error("Token audience does not match client ID");
     }
 
-    req.user = payload;
+    req.auth = payload;
+
     next();
   } catch (error) {
     res.status(401).json({ data: null, error: error });
